Xcode 10 beta 6

+++ iOS/PrivateFrameworks/iWorkImport.framework/TSCETableTransposedInfo.h
@@ -10,10 +10,10 @@ __attribute__((visibility("hidden")))
 @interface TSCETableTransposedInfo : NSObject
 {
     struct TSCERangeCoordinate _transposedBodyRange;
-    unsigned short _numberOfFooterRows;
+    unsigned int _numberOfFooterRows;
 }
 
-@property(nonatomic) unsigned short numberOfFooterRows; // @synthesize numberOfFooterRows=_numberOfFooterRows;
+@property(nonatomic) unsigned int numberOfFooterRows; // @synthesize numberOfFooterRows=_numberOfFooterRows;
 @property(nonatomic) struct TSCERangeCoordinate transposedBodyRange; // @synthesize transposedBodyRange=_transposedBodyRange;
 
 @end
