+++ iOS/PrivateFrameworks/iWorkImport.framework/TSTLayoutDynamicResizeInfoAccordion.h
@@ -18,7 +18,7 @@ __attribute__((visibility("hidden")))
 - (void)captureNewMinimumRowHeights:(id)arg1;
 - (double)applyResizeWidthFactor:(double)arg1;
 - (double)applyResizeHeightFactor:(double)arg1;
-- (double)p_applyAccordionResizeFactor:(double)arg1 region:(id)arg2 direction:(long long)arg3 totalOfCaptured:(double)arg4 count:(unsigned short)arg5 capturedArray:(double *)arg6 currentArray:(double *)arg7 minimumArray:(double *)arg8;
+- (double)p_applyAccordionResizeFactor:(double)arg1 region:(id)arg2 direction:(long long)arg3 totalOfCaptured:(double)arg4 count:(unsigned int)arg5 capturedArray:(double *)arg6 currentArray:(double *)arg7 minimumArray:(double *)arg8;
 - (id)initWithDynamicResizeInfo:(id)arg1;
 - (id)initWithMasterLayout:(id)arg1 columnRegion:(id)arg2 rowRegion:(id)arg3;
